+++ src/Repos/ChatFileRepo.ts
@@ -1,17 +1,15 @@
-import { FileStorage } from "../Services/FileStorage";
-import { ChatRepo } from "./Repo";
+import { DataStorage } from "../Services/DataStorage";
+import { ChatRepo } from "./ChatRepo";
 
-export const ChatFileRepo:() => ChatRepo = () => {
-    const reader = FileStorage();
-
+export const ChatFileRepo: (dataStorage: DataStorage) => ChatRepo = (dataStorage) => {
     const getChats = async (): Promise<number[]> => {
-        const chats = await reader.read("res/chats.json");
+        const chats = await dataStorage.read("res/chats.json");
         return JSON.parse(chats);
     }
 
     const setChats = async (chats: number[]) => {
         const json = JSON.stringify(chats);
-        await reader.write('res/chats.json', json);
+        await dataStorage.write('res/chats.json', json);
     }
 
     return { getChats, setChats };
